Project#2/MovieCatalogue/app/src/main/java/com/dicoding/moviecatalogue/ui/tvshow/TvShowViewModel.kt
package com.dicoding.moviecatalogue.ui.tvshow

import androidx.lifecycle.LiveData
import androidx.lifecycle.ViewModel
import androidx.paging.PagedList
import com.dicoding.moviecatalogue.data.source.CatalogueRepository
import com.dicoding.moviecatalogue.data.source.local.entity.TvShowEntity
import com.dicoding.moviecatalogue.vo.Resource

class TvShowViewModel(private val catalogueRepository: CatalogueRepository) : ViewModel() {

    fun getPopularTvShow(): LiveData<Resource<PagedList<TvShowEntity>>> =
        catalogueRepository.getPopularTvShow()

}
